requirements.txt
tensorflow
streamlit
numpy
pandas
pillow
matplotlib
scikit-learn
opencv-python-headless
